% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canceRbits.R
\docType{package}
\name{canceRbits}
\alias{canceRbits}
\title{canceRbits: Cancer genomics toolbox}
\description{
Wrapper and helper functions commmonly used by the Developmental
Cancer Genomics group at St. Anna Children's Cancer Research Institute.
}
\section{GSE functions}{

These are the main functions used for gene set enrichment:
\enumerate{
  \item \code{\link{cb_hyper}}
  \item \code{\link{cb_fgsea}}
}
}

\section{QC/filter functions}{

\enumerate{
  \item \code{\link{cb_filter_count_matrix}}
}
}

