% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{cb_celltypist}
\alias{cb_celltypist}
\title{Run celltypist}
\usage{
cb_celltypist(
  counts,
  ct_models = "Immune_All_High.pkl",
  ct_mode = "best_match",
  verbose = FALSE
)
}
\arguments{
\item{counts}{Count matrix with features as rows and cells as columns}

\item{ct_models}{Name(s) of celltypist model; default is 'Immune_All_High.pkl'}

\item{ct_mode}{Celltypist label prediction mode, either 'best_match' or 'prob_match'; default is 'best_match'}

\item{verbose}{Show or hide progress; default is FALSE}
}
\value{
A named list (one entry per model) of data frames of predictions
}
\description{
Run celltypist
}
\section{Details}{

This is a simple wrapper that writes the count matrix to a temporary file,
calls celltyper, reads the results.
Check \url{https://www.celltypist.org} for celltypist specific details.
For celltypist to find existing models, use \code{Sys.setenv(CELLTYPIST_FOLDER = path)},
where \code{path} must point to the directory that contains the data/models directories.

If \code{s} is a Seurat object, \code{ct_out <- celltypist(counts = s$RNA@counts)} would
run celltypist to get the annotations, \code{s <- AddMetaData(s, metadata = ct_out[[1]])}
would add them to the meta data of \code{s}
}

\examples{
\donttest{
# Coming soon
}

}
