% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplexing.R
\name{cb_demux_gmm}
\alias{cb_demux_gmm}
\title{De-multiplex}
\usage{
cb_demux_gmm(tag_counts, min_count = 30, pseudo_count = 10)
}
\arguments{
\item{tag_counts}{Count matrix with cells rows and barcodes (tags) as columns}

\item{min_count}{Minimum number of total tag counts required per cell}

\item{pseudo_count}{The pseudo-count that will be added to all observed counts before
log-transformation}
}
\value{
A data frame of results
The 'tag' column holds the final tag assignment and will match one of the
input column names, 'neg', or 'multi'
}
\description{
De-multiplex
}
\section{Details}{

Barcode-based demultiplexing using Gaussian finite mixture model (GMM) that
identifies per pair of barcodes four groups of cells:
G1) positive for barcode 1,
G2) positive for barcode 2,
G3) negative for both barcodes, and
G4) positive for both barcodes.
For each cell and barcode pair, we calculate the mean and relative differences
(difference over the mean) of log10-transformed counts (pseudo-count added).
In a first iteration, we use a 1-dimensional mixture model
(mclust package; MClustSSC method; parameters: modelNames=”E”, G = 3) with
relative differences as input and the following training data: the 50 cells
with highest/lowest relative difference as positives for G1 and G2, 50 cells
closest to the mean of G1 and G2 as G3. Based on the resulting classification
we generate synthetic G4 training data by sampling cells assigned to
G1 and G2 and combining their barcode counts (barcode 1 from G1 cells,
barcode 2 from G2 cells). These synthetic doublet cells are added to the
training data and a final 2-dimensional mixture model (parameters:
modelNames=”VVV”, G = 4) with relative difference and mean as input is used
for classification.
}

\examples{
\donttest{
# Coming soon
}

}
