% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{cb_enrichr_gsets}
\alias{cb_enrichr_gsets}
\title{Load local copy of EnrichR gene sets}
\usage{
cb_enrichr_gsets(filepath, verbose = TRUE)
}
\arguments{
\item{filepath}{Path to tsv file}

\item{verbose}{Boolean indicating whether to show messages (default is TRUE)}
}
\value{
A data frame of results
}
\description{
Load local copy of EnrichR gene sets
}
\section{Details}{

The input tsv files are meant to be in the format as provided
at \url{https://maayanlab.cloud/Enrichr/#libraries}.
}

\examples{
\donttest{
filepath <- paste0('/data_synology_rg3/cancerbits/users/',
                   'christoph_h/projects/gsea/enrichr_lib/',
                   'GO_Biological_Process_2021.tsv')
genesets <- cb_enrichr_gsets(filepath = filepath)
}

}
