% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{cb_filter_count_matrix}
\alias{cb_filter_count_matrix}
\title{Filter a scRNA-seq count matrix based on QC metrics}
\usage{
cb_filter_count_matrix(
  counts,
  percent_mito_th = 15,
  log_counts_z_th = c(-3, 3),
  feature_outlier_z_th = c(-5, 5),
  min_features = 300,
  mito_pattern = "^MT[^0-9a-zA-Z]+",
  return_seurat = TRUE,
  sample_id = NULL,
  verbose = TRUE
)
}
\arguments{
\item{counts}{Count matrix with features as rows and cells as columns}

\item{percent_mito_th}{Max percentage of mitochondrial counts; cell with more will be removed; default is 15}

\item{log_counts_z_th}{Z-score thresholds for sum of counts and sum of features
(these cell statistics will be log-transformed and scaled); default is \code{c(-3, 3)}}

\item{feature_outlier_z_th}{Z-score thresholds for the residuals of the UMI sum vs number
of features relationship; default is \code{c(-5, 5)}}

\item{min_features}{Min of detected features; cells with less will be removed; default is 300}

\item{mito_pattern}{Regex pattern used to identify mitochondrial features; default is \code{"^MT[^0-9a-zA-Z]+"}}

\item{return_seurat}{Boolean indicating whether to return the filtered matrix as Seurat object; default is TRUE}

\item{sample_id}{Optional string giving the sample name; will be used if return value is a Seurat object; default is NULL}

\item{verbose}{Boolean indicating whether to show messages; default is TRUE}
}
\value{
A list with components: 1) filtered object, 2) list of figures, 3) figure title
}
\description{
Filter a scRNA-seq count matrix based on QC metrics
}
\section{Details}{

Four filters are applied in succession:
1) percent mitochondrial reads
2) number of features
3) sum of counts
4) transcripts vs genes

Filters 1 to 3 are applied to the cell-level statistics; for 2) and 3) these are log10-transformed before Z-scores are calculated.
Filter 4) applies a loess fit to the relationship of the log10-transformed values of number of features as function of sum of counts, the residuals are then z-scored.
}

\examples{
\donttest{
filter_out <- canceRbits::cb_filter_count_matrix(sctransform::pbmc)
# show figures
patchwork::wrap_plots(filter_out$figures) + patchwork::plot_annotation(title = filter_out$fig_title)
# show filtered object
show(filter_out$filtered)
}

}
