% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{cb_hyper}
\alias{cb_hyper}
\title{Perform gene set enrichment using hypergeometric test}
\usage{
cb_hyper(
  signature,
  background,
  genesets,
  min_size = 9,
  max_size = 500,
  collapse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{signature}{Character vector of signature genes (the foreground set)}

\item{background}{Character vector of genes (the background set)}

\item{genesets}{List of character vectors of pre-defined gene sets (e.g. pathways)}

\item{min_size}{Minimum size of a gene set to be tested (after intersection with background) (default is 9)}

\item{max_size}{Maximum size of a gene set to be tested (after intersection with background) (default is 500)}

\item{collapse}{Boolean indicating whether in the output lower ranking gene sets should be removed if they are 99\% included in a higher ranking one (default is TRUE)}

\item{verbose}{Boolean indicating whether to show messages (default is TRUE)}
}
\value{
A data frame of results
}
\description{
Perform gene set enrichment using hypergeometric test
}
\section{Details}{

This is a light wrapper around hypeR::hypeR. This functions filters the gene sets
given the background and makes sure that there is empty output if there is nothing to be tested.
Output is ordered based on p-value. Lower ranking gene sets that are 99\% contained in higher ranking ones can be omitted.
}

\section{Notes}{

The background set should contain all (and only those) genes that were tested/considered to generate the signature gene set (aka foreground).

To load local copies of pathways from EnrichR, check out the helper
function \code{\link{cb_enrichr_gsets}}.
To load them from the web check out \code{\link[hypeR:enrichr_gsets]{hypeR::enrichr_gsets}}.
}

\section{TODO}{

Add documentation and more options to sanitize the geneset labels.
}

\examples{
\donttest{
background <- LETTERS
genesets <- split(LETTERS, f = round((1:length(LETTERS)) / 4))
names(genesets) <- paste0('gene_sets_', 1:length(genesets))
signature <- c('N', 'O', 'V', 'W', 'X')
cb_hyper(signature = signature,
         background = background,
         genesets = genesets, min_size = 2,
         collapse = FALSE, verbose = FALSE)
}

}
