% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_wrapper.R
\name{cb_infercnv_report}
\alias{cb_infercnv_report}
\title{Create an inferCNV report}
\usage{
cb_infercnv_report(
  counts,
  conditions,
  out_report_path,
  out_dir = NULL,
  ref_conditions = NULL,
  gene_order_file = NULL,
  num_threads = 1,
  convert = FALSE,
  cleanup = TRUE,
  ...
)
}
\arguments{
\item{counts}{Count matrix}

\item{conditions}{The cell conditions, or clusters/group labels; its length must match the number of columns of the count matrix}

\item{out_report_path}{Path of the resulting report}

\item{out_dir}{The output directory for inferCNV, will be created if it does not exist; the default is NULL and a temporary directory will be used}

\item{ref_conditions}{The conditions that make up the reference (e.g. wildtype); default is NULL
and the mean of all cells is used as reference}

\item{gene_order_file}{The file with gene to location mappings; default is NULL and it will be downloaded from
https://data.broadinstitute.org/Trinity/CTAT/cnv/gencode_v19_gene_pos.txt}

\item{num_threads}{The number of CPU threads inferCNV will use; default is 1}

\item{convert}{Boolean, if TRUE infercnv score files are converted to .Rds in the output directory; default is FALSE}

\item{cleanup}{Boolean, if TRUE remove intermediate infercnv output files; default is TRUE}

\item{...}{parameters passed to rmarkdown::render, e.g. quiet = TRUE}
}
\description{
Create an inferCNV report
}
\section{Details}{

See cb_run_infercnv for details
}

