% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_workflow.R
\name{cb_load_counts}
\alias{cb_load_counts}
\title{Load count matrix given file}
\usage{
cb_load_counts(path, modality = "Gene Expression")
}
\arguments{
\item{path}{File path (h5 or rds)}

\item{modality}{Slot to use in case file has multiple count matrices}
}
\value{
Sparse count matrix
}
\description{
Load count matrix given file
}
\section{Details}{

Looks at file extension to use the appropriate function to load the data.
Supported at the moment: H5, HDF5, RDS (capitalization does not matter)
}

