% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_workflow.R
\name{cb_seurat_pipeline}
\alias{cb_seurat_pipeline}
\title{Run a Seurat workflow}
\usage{
cb_seurat_pipeline(x, max_pc = 15, cluster_res = 0.8, verbose = FALSE)
}
\arguments{
\item{x}{Sparse matrix or Seurat object}

\item{max_pc}{The number of principal components to use; default is 15}

\item{cluster_res}{Resolution parameter used with FindClusters; default is 0.8}

\item{verbose}{Boolean indicating whether to print progress; is passed on
to most functions used internally; default is FALSE}
}
\value{
A list with items
\itemize{
\item s - the final Seurat object
\item figures - a list of figures
\item fig_title - a figure title
}
}
\description{
Run a Seurat workflow
}
\section{Details}{

Performs a fairly standard Seurat analysis workflow including
\itemize{
\item SCTransform
\item RunPCA
\item RunUMAP
\item FindNeighbors
\item FindClusters
}

It will also find positive markers of the clusters and save the results in
object$RNA@misc$markers
}

\section{TODO}{

Expose more of the parameters that are currently hard coded. Add examples.
}

\examples{
\donttest{
# Coming soon
}
}
