% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplexing.R
\name{demux_one_pair}
\alias{demux_one_pair}
\title{Demultiplex one pair of barcode counts}
\usage{
demux_one_pair(
  x,
  y,
  N = 50,
  class_names = c("neg", "x", "y", "multi"),
  label_levels = NULL
)
}
\arguments{
\item{x}{log-counts of barcode 1}

\item{y}{log-counts of barcode 2}

\item{N}{Number of training cells to use for semi-supervised classification}

\item{class_names}{The names of the output classes}

\item{label_levels}{If no NULL then the output will be a factor with these levels}
}
\value{
Character vector or factor of demultiplex classes
}
\description{
Demultiplex one pair of barcode counts
}
\keyword{internal}
