% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplexing.R
\name{cb_demux_gmm}
\alias{cb_demux_gmm}
\title{De-multiplex}
\usage{
cb_demux_gmm(tag_counts, min_count = 30, pseudo_count = 10)
}
\arguments{
\item{tag_counts}{Count matrix with cells rows and barcodes (tags) as columns}

\item{min_count}{Minimum number of total tag counts required per cell}

\item{pseudo_count}{The pseudo-count that will be added to all observed counts before
log-transformation}
}
\value{
A data frame of results
The 'tag' column holds the final tag assignment and will match one of the
input column names, 'neg', or 'multi'
}
\description{
De-multiplex
}
\section{Details}{

}

\examples{
\donttest{
# Coming soon
}

}
