% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_wrapper.R
\name{cb_single_sample_report}
\alias{cb_single_sample_report}
\title{Create a single sample report for scRNA-seq data}
\usage{
cb_single_sample_report(
  sample_counts,
  out_report_path,
  sample_name = "foo",
  out_rds_path = NULL,
  return_seurat = FALSE,
  seurat_max_pc = 15,
  seurat_cluster_res = 0.3,
  seurat_metric = "manhattan",
  seurat_k_param = 20,
  seurat_n_neighbors = 40,
  ...
)
}
\arguments{
\item{sample_counts}{Either a sparse count matrix, a path to h5 file, or a path to an rds file}

\item{out_report_path}{Path of the resulting report}

\item{sample_name}{Name to use for this sample; default is foo}

\item{out_rds_path}{Path of the resulting Seurat object (Rds file); default is NULL
and no file will. be written}

\item{return_seurat}{Boolean indicating whether to return the final seurat object;
default is FALSE and will return NULL invisibly}

\item{seurat_max_pc}{The number of principal components to use; default is 15}

\item{seurat_cluster_res}{Resolution parameter used with FindClusters; default is 0.3}

\item{seurat_metric}{Distance metric to use for FindNeighbors and RunUMAP; default is manhattan}

\item{seurat_k_param}{k.param parameter for FindNeighbors; default is 20}

\item{seurat_n_neighbors}{n.neighbors parameter for RunUMAP; default is 40}

\item{...}{parameters passed to rmarkdown::render, e.g. quiet = TRUE}
}
\description{
Create a single sample report for scRNA-seq data
}
\section{Details}{

Uses the single_sample_overview rmarkdown template
}

