% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_wrapper.R
\name{cb_single_sample_report}
\alias{cb_single_sample_report}
\title{Create a single sample report for scRNA-seq data}
\usage{
cb_single_sample_report(
  sample_counts,
  out_report_path,
  sample_name = "foo",
  out_rds_path = NULL,
  return_seurat = FALSE,
  ...
)
}
\arguments{
\item{sample_counts}{Either a sparse count matrix, a path to h5 file, or a path to an rds file}

\item{out_report_path}{Path of the resulting report}

\item{sample_name}{Name to use for this sample; default is foo}

\item{out_rds_path}{Path of the resulting Seurat object (Rds file); default is NULL
and no file will. be written}

\item{return_seurat}{Boolean indicating whether to return the final seurat object;
default is FALSE and will return NULL invisibly}

\item{...}{parameters passed to rmarkdown::render, e.g. quiet = TRUE}
}
\description{
Create a single sample report for scRNA-seq data
}
\section{Details}{

Uses the single_sample_overview rmarkdown template
}

